feat(client): rebalance dock, promote AUTO to header, merge dashboard title row
Live Performance moves from the More menu into the dock's left cluster
(5 items now balance the right cluster + avatar). AUTO focus-slot filler
moves from More into GameHeader as a cyan outline button (same shared-
service logic, toasts, spinner). Dashboard title block rises onto the
vitals row on lg+ per dashboard-dock.html; width capped so vitals stay
clickable.
